# src/Ims.Case.Api/wwwroot/docs/case-v1p0-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Competencies and Academic Standards Exchange (CASE) Service OpenAPI (JSON) Definition",
    "description": "The Competencies and Academic Standards Exchange (CASE) Service enables the exchange of data between a Competency Records Service Provider and the consumers of the associated data. This service has been described using the IMS Model Driven Specification development approach this being the Platform Specific Model (PSM) of the service.",
    "termsOfService": "http://www.opendefinition.org/licenses/odc-by",
    "contact": {
      "name": "Brad R. Marshall ",
      "email": "bmarshall@sped.mobi"
    },
    "license": {
      "name": "Open Data Commons Attribution License (ODCAL)",
      "url": "http://www.opendefinition.org/licenses/odc-by"
    },
    "version": "v1p0"
  },
  "servers": [
    {
      "url": "https://teks-api.texasgateway.org/ims/case/v1p0",
      "description": "TEA Texas Gateway Server"
    }
  ],
...
  "paths": {
...
    "/CFAssociations/{sourcedId}": {
      "get": {
        "tags": [
          "AssociationsManager"
        ],
        "summary": "The REST read request message for the getCFAssociation() API call.",
        "description": "This is a request to the service provider to provide the information for the specific Competency Framework Association. If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFAssociation",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the CFAssociation to be supplied.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFAssociation from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFAssociation"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
    "/CFConcepts/{sourcedId}": {
      "get": {
        "tags": [
          "DefinitionsManager"
        ],
        "summary": "The REST read request message for the getCFConcept() API call.",
        "description": "This is a request to the Service Provider to provide the specified Competency Framework Concept and the set of children CFConcepts as identified by the hierarchy codes.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFConcept",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the Competency Framework Concept that is to be read from the service provider.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFConcept and child CFSConcepts from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFConceptSet"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CFAssociation": {
        "type": "object",
        "description": "This is the container for the data about the relationship between two CFDocuments or between two CFItems outside of the context of a CFPackage or CFItem.",
        "allOf": [
          {
            "$ref": "#/components/schemas/CFPckgAssociation"
          },
          {
            "type": "object",
            "properties": {
              "CFDocumentURI": {
                "$ref": "#/components/schemas/LinkURI"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "CFConcept": {
        "required": [
          "hierarchyCode",
          "identifier",
          "lastChangeDateTime",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "keywords": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "hierarchyCode": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "description": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "The container for the definition of a concept which is addressed by the competency framework."
      },
      "CFConceptSet": {
        "required": [
          "CFConcepts"
        ],
        "type": "object",
        "properties": {
          "CFConcepts": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CFConcept"
            }
          }
        },
        "additionalProperties": false,
        "description": "The container for the set of CFConcepts supplied in the response payload. The relationship between the CFConcepts is determined by the 'hierarchyCode'. The first CFConcept is that which has been specified in the call. The other CFConcepts are the set of children as determined by their place in the 'hierarchyCode' of the CFConcept."
      },
...
      "CFPckgAssociation": {
        "required": [
          "associationType",
          "destinationNodeURI",
          "identifier",
          "lastChangeDateTime",
          "originNodeURI",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "associationType": {
            "type": "string",
            "enum": [
              "isChildOf",
              "isPeerOf",
              "isPartOf",
              "exactMatchOf",
              "precedes",
              "isRelatedTo",
              "replacedBy",
              "exemplar",
              "hasSkillLevel"
            ]
          },
          "sequenceNumber": {
            "type": "integer",
            "description": "Model Primitive Datatype = Integer",
            "format": "int32"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "originNodeURI": {
            "$ref": "#/components/schemas/LinkGenURI"
          },
          "destinationNodeURI": {
            "$ref": "#/components/schemas/LinkGenURI"
          },
          "CFAssociationGroupingURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the data about the relationship between two CFDocuments or between two CFItems within the context of a CFPackage."
      },
...
      "LinkGenURI": {
        "required": [
          "identifier",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "identifier": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          }
        },
        "additionalProperties": false,
        "description": "A container for the information that is used to achieve the link data reference."
      },
      "LinkURI": {
        "required": [
          "identifier",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          }
        },
        "additionalProperties": false,
        "description": "A container for the information that is used to achieve the link data reference."
      },
...
      "UUID": {
        "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}",
        "type": "string",
        "description": "The data-type for establishing a Globally Unique Identifier (GUID). The form of the GUID is a Universally Unique Identifier (UUID) of 16 hexadecimal characters (lower case) in the format 8-4-4-4-12. All permitted versions (1-5) and variants (1-2) are supported.",
        "format": "guid"
      },
      "imsx_CodeMinor": {
        "required": [
          "imsx_codeMinorField"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMinorField": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/imsx_CodeMinorField"
            }
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the set of code minor status codes reported in the responses from the Service Provider."
      },
      "imsx_CodeMinorField": {
        "required": [
          "imsx_codeMinorFieldName",
          "imsx_codeMinorFieldValue"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMinorFieldName": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "imsx_codeMinorFieldValue": {
            "type": "string",
            "enum": [
              "fullsuccess",
              "invalid_sort_field",
              "invalid_selection_field",
              "forbidden",
              "unauthorisedrequest",
              "internal_server_error",
              "unknownobject",
              "server_busy",
              "invaliduuid"
            ]
          }
        },
        "additionalProperties": false,
        "description": "This is the container for a single code minor status code."
      },
      "imsx_StatusInfo": {
        "required": [
          "imsx_codeMajor",
          "imsx_severity"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMajor": {
            "type": "string",
            "enum": [
              "success",
              "processing",
              "failure",
              "unsupported"
            ]
          },
          "imsx_severity": {
            "type": "string",
            "enum": [
              "status",
              "warning",
              "error"
            ]
          },
          "imsx_description": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "imsx_codeMinor": {
            "$ref": "#/components/schemas/imsx_CodeMinor"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the status code and associated information returned within the HTTP messages received from the Service Provider. For the CASE service this object will only be returned to provide information about a failed request i.e. it will NOT be in the payload for a successful request. See Appendix B for further information on the interpretation of the information contained within this class"
      }
    }
  }
}